+++ lib/array.adt.h
@@ -12,7 +12,7 @@ void array_destroy(array_t *a);
 void array_append(array_t *a, const void *item, size_t siz);
 void *array_peek(array_t *a, int idx);
 void array_foreach(array_t *a, array_iterator_t *iterate);
-int array_size(array_t *a);
+size_t array_size(array_t *a);
 void array_sort(array_t *a, array_sorter_t *compare);
 
 #endif
